backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
google-cloud-storage
google-cloud-bigquery
google-cloud-secret-manager
google-cloud-logging
firebase-admin
pandas
statsmodels
scikit-learn
pydantic
sqlalchemy
psycopg2-binary
cloud-sql-python-connector
